Split providers schools features into files per action
